Tooltip refactor
